Remove data from CompiledCard, refactor routes to support

+++ server/src/routes/card-routes.ts
@@ -10,6 +10,7 @@ export async function respondWithCardForPath(
 ) {
   let {
     builder,
+    realms,
     cardRouter,
     params: { pathname },
   } = ctx;
@@ -24,20 +25,23 @@ export async function respondWithCardForPath(
     throw new NotFound(`No card defined for route ${pathname}`);
   }
 
+  let rawCard = await realms.getRawCard(url);
   let card = await builder.getCompiledCard(url);
-  ctx.body = await serializeCard(card, 'isolated');
+  ctx.body = await serializeCard(url, rawCard.data, card['isolated']);
   ctx.status = 200;
 }
 
 export async function getCard(ctx: RouterContext<any, CardStackContext>) {
   let {
     builder,
+    realms,
     params: { encodedCardURL: url },
   } = ctx;
 
   let format = getCardFormatFromRequest(ctx.query.format);
+  let rawCard = await realms.getRawCard(url);
   let card = await builder.getCompiledCard(url);
-  ctx.body = await serializeCard(card, format);
+  ctx.body = await serializeCard(url, rawCard.data, card[format]);
   ctx.status = 200;
 }
 
@@ -46,6 +50,7 @@ export async function createDataCard(
 ) {
   let {
     builder,
+    realms,
     request: { body },
     params: { parentCardURL, realmURL },
   } = ctx;
@@ -62,38 +67,50 @@ export async function createDataCard(
 
   let data = body.data as any;
 
-  let compiledCard = await builder.createDataCard(
-    realmURL,
-    parentCardURL,
-    data
-  );
+  let rawCard = await realms
+    .getRealm(realmURL)
+    .createDataCard(data.attributes, parentCardURL, data.id);
+
+  let compiledCard = await builder.getCompiledCard(rawCard.url);
 
   let format = getCardFormatFromRequest(ctx.query.format);
 
-  ctx.body = await serializeCard(compiledCard, format);
+  ctx.body = await serializeCard(
+    compiledCard.url,
+    rawCard.data,
+    compiledCard[format]
+  );
   ctx.status = 201;
 }
 
 export async function updateCard(ctx: RouterContext<any, CardStackContext>) {
   let {
     builder,
+    realms,
     request: { body },
     params: { encodedCardURL: url },
   } = ctx;
 
   let data = await deserialize(body);
-  let card = await builder.updateCardData(url, data.attributes);
+  let rawCard = await realms.updateCardData(url, data.attributes);
+
+  let card = await builder.getCompiledCard(url);
 
   // Question: Is it safe to assume the response should be isolated?
-  ctx.body = await serializeCard(card, 'isolated');
+  ctx.body = await serializeCard(url, rawCard.data, card['isolated']);
   ctx.status = 200;
 }
 
 export async function deleteCard(ctx: RouterContext<any, CardStackContext>) {
   let {
     builder,
+    realms,
     params: { encodedCardURL: url },
   } = ctx;
+
+  if (!realms.doesCardExist(url)) {
+    throw new NotFound(`Card ${url} does not exist`);
+  }
 
   builder.deleteCard(url);
